requirements.txt
torch>=1.12.1
transformers>=4.30.0
datasets>=2.8.0
peft>=0.4.0
trl>=0.4.0
PyYAML>=6.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.64.0
bitsandbytes
wandb